Refactor platform API imports to use shared module path
