+++ apps/electron/src/background.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import { createRoot } from "react-dom/client";
 import BackgroundComponent from "@/renderer/components/agent/BackgroundComponent";
-import { PlatformAPIProvider } from "@/main/infrastructure/platform-api";
+import { PlatformAPIProvider } from "@/renderer/platform-api";
 import { electronPlatformAPI } from "@/renderer/lib/electron-platform-api";
 import "@mcp_router/tailwind-config/base.css";
