+++ apps/electron/src/renderer/components/App.tsx
@@ -32,7 +32,7 @@ import {
   useUIStore,
   initializeStores,
 } from "../stores";
-import { usePlatformAPI } from "@/main/infrastructure/platform-api";
+import { usePlatformAPI } from "@/renderer/platform-api";
 import { IconProgress } from "@tabler/icons-react";
 import { postHogService } from "../services/posthog-service";
