+++ apps/electron/src/renderer/components/Sidebar.tsx
@@ -11,7 +11,7 @@ import {
 } from "@tabler/icons-react";
 import { useTranslation } from "react-i18next";
 import { useAuthStore, useWorkspaceStore } from "@/renderer/stores";
-import { usePlatformAPI } from "@/main/infrastructure/platform-api";
+import { usePlatformAPI } from "@/renderer/platform-api";
 // @ts-ignore
 import iconImage from "../../../../../public/images/icon/icon.png";
 import {
