+++ apps/electron/src/renderer/components/agent/AgentAuthGuard.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { useAuthStore } from "@/renderer/stores";
-import { usePlatformAPI } from "@/main/infrastructure/platform-api";
+import { usePlatformAPI } from "@/renderer/platform-api";
 import { LoginScreen } from "@/renderer/components/auth/LoginScreen";
 
 interface AgentAuthGuardProps {
