+++ apps/electron/src/renderer/components/agent/PackageManagerOverlay.tsx
@@ -16,7 +16,7 @@ import {
 } from "lucide-react";
 import { toast } from "sonner";
 import { useTranslation } from "react-i18next";
-import { usePlatformAPI } from "@/main/infrastructure/platform-api";
+import { usePlatformAPI } from "@/renderer/platform-api";
 
 interface PackageManagerOverlayProps {
   isOpen: boolean;
