+++ apps/electron/src/renderer/components/agent/use/AgentChat.tsx
@@ -9,7 +9,7 @@ import ChatInterface from "@/renderer/components/agent/create/ChatInterface";
 import ChatSessions from "./ChatSessions";
 import { isAgentConfigured } from "@/main/domain/agent/shared/agent-utils";
 import { useAgentStore } from "../../../stores";
-import { usePlatformAPI } from "@/main/infrastructure/platform-api";
+import { usePlatformAPI } from "@/renderer/platform-api";
 import { parseErrorMessage } from "@/renderer/utils/error-message-utils";
 
 /**
