+++ apps/electron/src/renderer/components/mcp/rules/Rules.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from "react";
 import { useTranslation } from "react-i18next";
-import { usePlatformAPI } from "@/main/infrastructure/platform-api";
+import { usePlatformAPI } from "@/renderer/platform-api";
 import { useWorkspaceStore } from "../../../stores";
 import {
   MCPDisplayRules,
