+++ apps/electron/src/renderer/components/mcp/server/server-details/ServerDetailsLocal.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from "react";
 import { MCPServer } from "@mcp_router/shared";
-import { usePlatformAPI } from "@/main/infrastructure/platform-api";
+import { usePlatformAPI } from "@/renderer/platform-api";
 import { useTranslation } from "react-i18next";
 import {
   Terminal,
