+++ apps/electron/src/renderer/platform-api/index.ts
@@ -1,0 +1,12 @@
+/**
+ * Platform API exports for Electron
+ *
+ * This module provides platform API utilities
+ * specifically for the Electron application
+ */
+
+// Platform API React context and provider (for backward compatibility)
+export { PlatformAPIProvider } from "./platform-api-context";
+
+// Export the store-based hook instead of the context-based one
+export { usePlatformAPI } from "@/renderer/lib/hooks/use-platform-api";
