+++ apps/electron/src/renderer/stores/server-store.ts
@@ -1,6 +1,5 @@
 import { create, StoreApi, UseBoundStore } from "zustand";
-import { MCPServer, MCPServerConfig, ServerState } from "@mcp_router/shared";
-import { PlatformAPI } from "@/main/infrastructure/platform-api";
+import { MCPServer, MCPServerConfig, ServerState, PlatformAPI } from "@mcp_router/shared";
 
 export interface ServerStoreState extends ServerState {
   // Actions
