+++ packages/shared/src/types/index.ts
@@ -14,7 +14,35 @@ export * from "./auth";
 // Re-export organized domain types
 export * from "./ui";
 export * from "./database";
-export * from "./platform-api";
+// Export platform-api types except LogEntry to avoid conflict
+export {
+  // Auth API
+  AuthAPI,
+  AuthStatus,
+  AuthProvider,
+  Unsubscribe,
+  // Server API
+  ServerAPI,
+  ServerStatus,
+  CreateServerInput,
+  // Agent API
+  AgentAPI,
+  // App API
+  AppAPI,
+  // Package API
+  PackageAPI,
+  // Settings API
+  SettingsAPI,
+  // Log API
+  LogAPI,
+  LogQueryOptions,
+  LogQueryResult,
+  // Workspace API
+  WorkspaceAPI,
+  // Main Platform API
+  PlatformAPI,
+} from "./platform-api";
+export { LogEntry as PlatformLogEntry } from "./platform-api";
 export * from "./mcp-apps";
 export * from "./utils";
 export * from "./cli";
